Improve timing accuracy and efficiency
Current timing code in monodroid-glue.cc uses the `gettimeofday` which provides
a most a microsecond accuracy, not enough for certain measurements.
Additionally, current code rounds it up to the millisecond boundery, making the
measurements very inaccurate. Also, `gettimeofday` is quite heavy in terms of
time spent in the syscall.

Replace `gettimeofday` with a call to `clock_gettime` when targetting Android or
Linux. This allows us to specify an alternative lock source, `CLOCK_MONOTONIC`,
which is very accurate and fast (not as fast as `CLOCK_MONOTONIC_COARSE` but
with higher resolution, thus better suited for our needs).

In addition, introduce a small collection of structures to be used for timing
instead of the currently used `long long` variables. Also introduced are the
`Util` class APIs to record the time (using the new `timing_period` structure
which has members to mark both the start and the end of the timing period - the
`start` and `end` fields, respectively) as well as a function to calculate the
difference between times stored in `time_period`. Structure representing the
difference supports reporting in seconds + milliseconds + nanoseconds.