[Xamarin.Android.Build.Tasks] Enable Aapt2 by default (#2381)
Enable `aapt2` as the default resource processor. 

Use of `aapt2` can be disabled by setting `$(AndroidUseAapt2)`=False
in the `.csproj`.

lso fixes a small issue with compiling library resources.
The scenario is as follows:

 1. Library project contains resources:

        Resource\drawable\foo.png
        Resource\layout\foo.xml

 2. Main app project references a library in (1).

 3. Build the App project.  This works as expected.

 4. Update `Resource\layout\foo.xml` in the Library project.

 5. Build again.

    This build does *not* refresh the `.flata` file for the library,
    as `<CollectNonEmptyDirectories/>` only looked for the first file
    in the library `res` directory, which was used as an input to the
    `_ConvertLibraryResourcesCases` and
    `_CompileAndroidLibraryResources` targets.
    
    If the Library project *isn't* the first file -- which is often
    the case -- then those targets won't be executed, and the
    `.flata` file won't be updated.

Instead of using the first file, use the library `.stamp` files as
the inputs.  This is because when we do extract new/updated files
from the library that `.stamp` file will be modified.  By using the
`.stamp` files for the inputs we know that those targets will be run.

Finally, update the `ManagedResourceParser` output so that nested
types are *always* generated.  This is what `aapt` output generates,
and skipping nested types -- as was previously done -- 
[resulted in API breakage][0], as the
`Xamarin.Android.NUnitLite.Resource.Attribute` type was not generated.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder/4522/API_20Compatibility_20Checks/