[Xamarin.Android.Build.Tasks] set $(AndroidSdkDirectory) if blank
Commit 6e7be95b, applied to `Xamarin.Android.Binding.targets` too.

As with 6e7be95b, a downstream project in `monodroid` imports
`Xamarin.Android.Bindings.targets` and uses `$(JavaSdkDirectory)`
without otherwise setting it, resulting in build failures when it
attempts to execute `/bin/javac`.

As with 6e7be95b, set the `$(JavaSdkPath)`, `$(AndroidNdkPath)`, and
`$(AndroidSdkPath)` MSBuild properties if they're not already set.