[monodroid] Improve timing accuracy and efficiency (#2467)
Current timing code in `monodroid-glue.cc` uses **gettimeofday**(2)
which provides at most microsecond accuracy, which isn't enough for
certain measurements.  Additionally, current code rounds it up to the
millisecond boundary, making the measurements very inaccurate.
Finally, `gettimeofday` can be quite heavy in terms of time spent in
the syscall.

Replace `gettimeofday()` with a call to **clock_gettime**(3) when
running on Android or Linux.  This allows us to specify an alternate
clock source, `CLOCK_MONOTONIC`, which is very accurate and fast (not
as fast as `CLOCK_MONOTONIC_COARSE` but with higher resolution, and
thus better suited for our needs).

In addition, introduce a small collection of structures to be used
for timing instead of the currently used `long long` variables.