[Xamarin.Android.Build.Tasks] Fix Aapt2 LowerCasing when using library resources.
We found an issue where if an application layout file
was referencing a library project resource it was NOT
being lowercased. This would result in the following
error

	error APT0000: resource drawable/Case_Check not found

The reason for this was the fact that we were only passing in
the `$(MonoAndroidResDirIntermediate)` directory to
`ConvertResourcesCases`. The Task was a bit of a strange one in
that it would use the same property for both the list of directories
it needed to alter AND the ones it needed to use for reference.
This puts us in a bit of a pickle because in the case of `aapt2`
we only want to work on that ONE directory.

So this PR introduces an additional parameter to the task

	`AdditionalResourceDirectories`

The purpose of this parameter is to provide a list of directories
to check for resources, but NOT alter or change. They are for
reference only. This will allow use to process a single
directory without being forces to process a bunch of others.
Those "other" directories are already processed by the
`_ConvertLibraryResourcesCases` target. This is part of how
`aapt2` work in that we are creating the `.flata` archives for
each library project as well as a seperate one for the main app.
This allows us to do faster incremental builds since the library
resources should rarely change.