[Xamarin.Android.Build.Tasks] Aapt2 LowerCasing & lib resources (#2483)
We found an issue where if an application layout file was referencing
a library project resource, the library project resource name was NOT
being lowercased.  This would result in the following error:

	error APT0000: resource drawable/Case_Check not found

The reason for this was the fact that we were only passing in the
`$(MonoAndroidResDirIntermediate)` directory to
`<ConvertResourcesCases/>`.  The Task was a bit of a strange one in
that it would use the same property for both the list of directories
it needed to alter *and* the ones it needed to use for reference.
This puts us in a bit of a pickle because in the case of `aapt2` we
only want to work on that ONE directory.

Add a new `ConvertResourcesCases.AdditionalResourceDirectories`
property, which is an `ITaskItem[]` list of directories to check for
resources, but *not* alter or change.  They are for reference only.
This will allow use to process a single directory without being
forced to process a bunch of others.  Those "other" directories are
already processed by the `_ConvertLibraryResourcesCases` target.
This is part of how `aapt2` work in that we are creating the
`.flata` archives for each library project as well as a separate one
for the main app.  This allows us to do faster incremental builds
since the library resources should rarely change.