[samples] use $(DebugType) portable
In 1fb7983, a new `HelloLibrary.csproj` was added, but it had some
settings that cause `pdb2mdb` to run:

    <DebugType>full</DebugType>
    ...
    <DebugType>pdbonly</DebugType>

I switched the "Hello World" projects to use `$(DebugType)` `portable`,
so that portable pdb files are always generated as part of their
build.

This way when profiling builds on Windows, I don't have to account for
the `<ConvertDebuggingFiles/>` MSBuild task taking time.