Add fields as needed for API-level-specific properties
As I understand it, Xamarin.Android.Build.Tasks.csproj follows a
convention where it currently specifies only `ANDROID_24` in
`$(DefineConstants)` and not any other API levels.  So the version of
the `ActivityAttribute` class that is part of the build tasks does not
include all of the available public properties for all API levels.
Instead, the additional partial class in ApplicationAttribute.Partial.cs
adds private fields for any absent public properties, and the build
tasks use those private fields to store and work with the values.

Add the private fields as needed for the new properties and alphabetize
all the private fields.

Make `LockTaskMode` a `string` rather than an
`Android.App.LockTaskMode`.  Although the name of
`Android.App.LockTaskMode` matches the new property name, the valid
values are different.  The valid values for the `LockTaskMode` property
correspond to [a set of fields in `android.content.pm.ActivityInfo` that
are marked `@hide`][0], so those values are not currently included in
the Xamarin.Android bindings.  For now, users will need to enter one of
the valid string values by hand instead.

Add an `[Obsolete]` attribute on `ShowOnLockScreen` because it is marked
`@deprecated` in sdk/platforms/*/data/res/values/attrs_manifest.xml

Add properties for the new `android:showWhenLocked` and
`android:turnScreenOn` attributes that are part of API level 27.

Add a `float` entry to `ValueConverters` so that the `MaxAspectRatio`
property, which was added a while ago, can now be used as expected.

Adjust the minimum API level to 11 for `Immersive` based on the output
from [build-tools/manifest-attribute-codegen][1] so that it will be a
tiny bit easier to do future comparisons using the output from that
tool.

Add a test to make sure all the old and new properties continue to
generate the expected XML attributes in the future.  This test does not
attempt cover all possible values for properties that take `enum`s.

The current full list of XML attributes found by
build-tools/manifest-attribute-codegen contains a few others that are
intentionally excluded from this commit:

  * `primaryUserOnly` is no longer present after API level 23, was
    always marked `@hide` in attrs_manifest.xml before that, and is not
    offered by autocompletion in Android Studio.
  * `systemUserOnly` is marked `@hide` in attrs_manifest.xml and is not
    offered by autocompletion in Android Studio.
  * `alwaysFocusable` is marked `@hide` in attrs_manifest.xml and is not
    offered by autocompletion in Android Studio.
  * `splitName` is used for Android Instant Apps, which are [not
    currently supported][2].

[0]: https://android.googlesource.com/platform/frameworks/base/+/pie-release/core/java/android/content/pm/ActivityInfo.java#915
[1]: https://github.com/xamarin/xamarin-android/tree/master/build-tools/manifest-attribute-codegen
[2]: https://github.com/xamarin/xamarin-android/blob/master/Documentation/project-docs/InstantApps.md