[tests] Clear the adb log ater we collect the logcat output
That way it is easier to find a problem in the logcat output as it
will only contain the information relevant to the particular test run.

As we clear it after the run, the first test logcat on a freshly
started emulator will also contain the startup info in case it is
needed.