[Xamarin.Android.Build.Tasks] fix Inputs/Outputs for _LinkAssemblies*
Downstream in monodroid, we are seeing an issue with "Fast Deployment":

1. An incremental build occurs; one or more assemblies change.
2. We still seem to deploy *every* assembly.

This started happening in 539954c, since I corrected the directory
that `_LinkAssembliesNoShrink` was touching files in.

`_LinkAssembliesNoShrink` was setup to:

1. Its `Outputs` were assemblies produced by the linker.
2. A `<Touch/>` command was updating timestamps of these assemblies.

Refactoring this target to more closely match `_LinkAssembliesShrink`
actually seems more correct here:

1. `Inputs` should include `$(_AndroidBuildPropertiesCache)`. Whoops!
2. `Outputs` should be the `$(_AndroidLinkFlag)` file.
3. We can do away with the `@(_LinkAssembliesNoShrinkFiles)` item
   group, and just directly add to `@(FileWrites)`.

Since we have improvements earlier in the build chain, I don't think
we need to `<Touch/>` the assemblies output from the linker at all
here.

I made similar adjustments in `_LinkAssembliesShrink` so it matches
`_LinkAssembliesNoShrink`.