[build] Makefile should run /t:Restore after dependencies.csproj (#2502)
We want the Mono upgrade to run *before* we run `/t:Restore`.

In one case downstream in monodroid, we encountered:

	error MSB4244: The SDK resolver assembly "/Library/Frameworks/Mono.framework/Versions/5.16.0/lib/mono/msbuild/15.0/bin/SdkResolvers/Microsoft.Build.NuGetSdkResolver/..\..\Microsoft.Build.NuGetSdkResolver.dll" could not be loaded. Invalid Image

This error normally means the file didn't exist, although it *does*
exist -- if we replace `\` with `/`.  Manually running the ~same
MSBuild command with `MONO_IOMAP=all` set *works*:

	MONO_IOMAP=all msbuild /t:Restore ...

Thus, the problem is that we're using the *wrong* system mono to run
the `/t:Restore` target.

The solution is to move the
`Xamarin.Forms.Performance.Integration.csproj` restoration to happen
*after* `build-tools/dependencies/dependencies.cproj` is built, which
will ensure that the appropriate system mono is installed.