[Mono.Android] Add more manifest attributes for activities (#2109)
Fixes: https://github.com/xamarin/xamarin-android/issues/1891 

The `Android.App.ActivityAttribute` custom attribute is used to
generate the [`<activity/>` element within `AndroidManifest.xml`][1].

There is an important philosophical question: should
`ActivityAttribute` provide property for *every known* `<activity/>`
attribute?  Or just the documented ones?

For quite some time, our stance has been that `ActivityAttribute`
should only provide the *documented* attributes.  However, there is a
flaw with this argument: documented *where*?  It turns out that
several attributes are *not* documented with the `<activity/>`
element documentation, but instead in "feature-specific"
documentation.  For example, the `//activity/@android:enableVrMode`
attribute is documented in the [Android Manifest for Google VR][4]
documentation and is *not* mentioned in the `AndroidManifest.xml`
element documentation!

Furthermore, we don't know where all we should be looking for the
feature-specific documentation!

Meanwhile, the Android SDK *itself* contains information about what
attributes are supported on a given API level: the
`$(AndroidSdkDirectory)\platforms\android-*\data\res\values\attrs_manifest.xml`
files specify which attributes are valid on elements, and
[build-tools/manifest-attribute-codegen][0] is able to read this
information and turn it into a more usable form.

Given our inability to keep on top of the various forms of Android
SDK feature documentation, we have changed our answer to the
philosophical question: we should expose *all* of the element
attributes as custom attribute properties, *not* just the documented
attributes.

With this background in mind, begin (manually) binding the
`<activity/>` attributes which haven't already been bound within
`ActivityAttribute`.  We will consider adding tooling to *generate*
the custom attributes based on the Android SDK data in a future PR.

The current full list of XML attributes found by
`build-tools/manifest-attribute-codegen` contains a few other
attributes that are intentionally *excluded* from this commit:

  * `primaryUserOnly` is no longer present after API level 23, was
    always marked `@hide` in `attrs_manifest.xml` before that, and is
    not offered by autocompletion in Android Studio.
  * `systemUserOnly` is marked `@hide` in `attrs_manifest.xml` and is
    not offered by autocompletion in Android Studio.
  * `alwaysFocusable` is marked `@hide` in `attrs_manifest.xml` and is
    not offered by autocompletion in Android Studio.
  * `splitName` is used for Android Instant Apps, which are
    [not currently supported][2].

Additionally:

  * Add support for `float` property values.
  * Add a test to make sure all the old and new properties continue
    to generate the expected XML attributes in the future.
    This test does *not* attempt cover all possible values for
    properties that take `enum`s.
  * `ShowOnLockScreen` is marked `[Obsolete]` because it is marked
    `@deprecated` in `sdk/platforms/*/data/res/values/attrs_manifest.xml`
  * `ActivityAttribute.LockTaskMode` is bound as a `string` and *not*
     as `Android.App.LockTaskMode` because the valid values are
     different.  The valid values for the `LockTaskMode` property
     correspond to a set of fields in
     [`android.content.pm.ActivityInfo` that are marked `@hide`][3],
     so those values are not currently included in the
     Xamarin.Android bindings.  For now, users will need to enter one
     of the valid string values by hand instead.

[0]: https://github.com/xamarin/xamarin-android/tree/master/build-tools/manifest-attribute-codegen
[1]: https://developer.android.com/guide/topics/manifest/activity-element
[2]: https://github.com/xamarin/xamarin-android/blob/master/Documentation/project-docs/InstantApps.md
[3]: https://android.googlesource.com/platform/frameworks/base/+/pie-release/core/java/android/content/pm/ActivityInfo.java#915
[4]: https://developers.google.com/vr/reference/vr-manifest