[Xamarin.Android.Build.Tasks] fix <Aot/> cancellation
Fixes: http://feedback.devdiv.io/346386

It was reported that cancelling a build while AOT was running, would
not cancel until AOT finished. Testing in VS Windows 15.9, I could
reproduce this behavior.

Reviewing the `<Aot/>` MSBuild task, it had use of a
`CancellationTokenSource` that killed the AOT processes if cancelled.
However, this did not work as intended, since the
`CancellationTokenSource` is a local variable and the `Cancel` method
can't call it.

When MSBuild tasks are cancelled, if they implement `ICancelableTask`,
the `Cancel` method will be called. `AsyncTask` has a proper
implementation of this, as well as a `Token` property for use with the
TPL. I removed the `CancellationTokenSource` from `<Aot/>`, in favor
of what is already implemented in `AsyncTask`.

I suspect that cancellation was added to `AsyncTask` at some point,
and we just never switched the `<Aot/>` MSBuild task to use it.