[Xamarin.Android.Build.Tasks] _ValidateResourceCache incremental build issue (#2520)
If the `_ValidateResourceCache` MSBuild target deems the cache file
invalid:

    <Delete Files="$(_AndroidResourcePathsCache)" Condition=" '$(_IsResourceCacheValid)' == 'False' " />

The problem is that the `_BuildAdditionalResourcesCache` will be
skipped!

    <Target Name="_BuildAdditionalResourcesCache"
        Inputs="@(ReferencePath);@(ReferenceDependencyPaths);$(MSBuildProjectFullPath);$(NugetPackagesConfig);$(_AndroidBuildPropertiesCache)"
        Outputs="$(_AndroidStampDirectory)_BuildAdditionalResourcesCache.stamp"

In eca09c89, I fixed a problem with `_BuildAdditionalResourcesCache`
running on every build by introducing a stamp file.

Since the `Outputs` are now the stamp file, we need to *also* delete
the stamp file when `'$(_IsResourceCacheValid)' == 'False'`. We had a
test checking this exact problem, but it wasn't verifying that
`_BuildAdditionalResourcesCache` wasn't skipped.