[Xamarin.Android.Build.Tasks] Remove EmbeddedNativeLibrary from Mono.Data.Sqlite. (#2417)
Fixes: https://github.com/xamarin/xamarin-android/issues/2407

We currently embed the required native libraries for
`Mono.Data.Sqlite.dll` and `Mono.Posix.dll` into the assembly as
an `@(EmbeddedResource)` (within `__AndroidNativeLibraries__.zip`).
This is then extracted as part of the build process into the `.apk`.
This is the only actual change we have from the vanilla versions of
both of these assemblies.

This commit removes the `@(EmbeddedResource)` items from these
assemblies.  It also updates `src/sqlite-xamarin` to place its
outputs into these directories:

	`lib\xamarin-android\xbuild\Xamarin\Android\lib\{arch}\`

This will make sure we install them as part of the various installers
we have.

Also update the MSBuild targets to check to see if these assemblies
are referenced and include the required `.so` files when they are.