[Xamarin.Android.Tools.BootstrapTasks] Include Logcat in errors
Context: https://jenkins.mono-project.com/job/xamarin-android-pr-builder-release/189/

Commit c38c58e9 updated the `<RenameTestCases/>` task so that if the
NUnit XML output file couldn't be found, we would instead *generate* a
"reasonable" NUnit XML file containing an error, so that Jenkins would
nicely show that *something* went wrong.

The result is that it works, but isn't "nice": the Test Result failure
is a *filesystem path*, e.g.

	/Users/builder/jenkins/workspace/xamarin-android-pr-builder-release/xamarin-android/build-tools/scripts/../../TestResult-Xamarin.Android.EmbeddedDSO_Test.nunit-Release-Aot.xml.TestResult-Xamarin.Android.EmbeddedDSO_Test.nunit-Release-Aot.xml

(That's *literally* the name of a "failing test.")

Fix this by constructing a reasonable alternative: use the
`SourceFile` basename -- without the `TestResult-` prefix and without
any extension -- as the `//test-suite/@name` value, and use
`Possible Crash / {Configuration}` as the `//test-case/@name` value.
For the above failure, this would result in this test failing:

	Xamarin.Android.EmbeddedDSO_Test.nunit.Possible Crash / Release-Aot

Additionally, the [*contents* of the generated file][0] aren't entirely
helpful: it's just a `FileNotFoundException`, because the expected
NUnit XML input doesn't exist.

Fix this by checking for `adb logcat` output within the
`{DestinationFolder}/tests` directory, with a filename matching
`logcat-{Configuration}-{testSuiteName}.txt`.  If this file exists,
the contents of the file are included into the NUnit XML file.

Unfortunately, the `logcat*.txt` file may contain non-textual
data (?!) which can't go through `XDocument.Save()` -- it throws an
`ArgumentException` from `XmlUtf8RawTextWriter.WriteCDataSection()` --
so "escape" blocks of binary data by emitting a `[!binary-block ...]`
text run which contains the ASCII representation of the invalid bytes.

[0]: https://jenkins.mono-project.com/job/xamarin-android-pr-builder-release/189/testReport/junit/_Users_builder_jenkins_workspace_xamarin-android-pr-builder-release_xamarin-android_build-tools_scripts_.._.._TestResult-Xamarin.Android.EmbeddedDSO_Test.nunit-Release-Aot/xml/TestResult_Xamarin_Android_EmbeddedDSO_Test_nunit_Release_Aot_xml/