[build-tools] Fix the RunUI task to wait
Fixes https://github.com/xamarin/xamarin-android/issues/2552

Use override in the right place. Probably a typo, which made the
logcat output unusable, because the `AfterCommand` was never called
and we gathered the logcat directly after launching the
activity. Instead of waiting before collecting the logcat.

My wild guess is that older compiler versions worked differenly and so
it worked for some time.