[Xamarin.Android.Build.Tasks] skip _SetLatestTargetFrameworkVersionForPackageReference (#2554)
During a NuGet restore:

	msbuild foo.android.csproj /t:Restore

The `_SetLatestTargetFrameworkVersionForPackageReference` MSBuild
target triggers `_SetLatestTargetFrameworkVersion` and some tasks
to run:

  - `<ResolveSdks/>`
  - `<ValidateJavaVersion/>`
  - `<ResolveAndroidTooling/>`

We don't really need these to run at all during `Restore`, unless
`$(AndroidUseLatestPlatformSdk)` is `True`.  In that case we need to
calculate what the final `$(TargetFrameworkVersion)` will be.

Optionally running this target saved ~600ms during a NuGet restore of
a single Xamarin.Android project:

	602 ms  _SetLatestTargetFrameworkVersion           1 calls

With `_SetLatestTargetFrameworkVersionForPackageReference` only
happening when `$(AndroidUseLatestPlatformSdk)`=True -- which should
*rarely* be the case now -- the `BuildTest.ValidateUseLatestAndroid()`
test now needs to be updated, as there are now fewer instances of
`TargetFrameworkVersion` within the build log.