[Xamarin.Android.Build.Tests] Fix Repeat Build Tests
PR #2515 highlighted an issue with a few of our
repeat build tests. Tests were failing because
the `_Sign` target was being skipped. Looking
at the build logs, `BuildApk` was being called.
As was `_CreateBaseApk`, so what is going on?

So it turns out that in `BuildApk` we use the
`CopyZipIfChanged` method to move the temp
zip file over to the actual one. In this case
... the zip was identical! So the "new" zip
was not copied.. and so the `_Sign` task did
not need to run.

How can the zip be the same? It had an updated
file? Well wrong.. we just touched the timestamp.
In which case the zip ends up with the exact same
CRC values for each file. We use the CRC to detect
changes.. so no changes. So it turns out our build
system was behaving as expected.. it was just our
test was invalid.

So the fix here is to update the tests to make sure we
do change a file. This means the CRC's will change and
the targets will run.